Move vpp functions to mysql/vpp.go
